Factor out the Slice class into utils
Fixes issue #15
