+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/io/AbstractIO.java
@@ -11,6 +11,7 @@ import java.io.OutputStream;
 
 import uk.ac.manchester.spinnaker.processes.FillProcess;
 import uk.ac.manchester.spinnaker.processes.Process.Exception;
+import uk.ac.manchester.spinnaker.utils.Slice;
 
 /**
  * Presents a view of an entity (SpiNNaker memory or a file on disk) as
@@ -236,60 +237,6 @@ public interface AbstractIO extends AutoCloseable {
 		CUR,
 		/** Seek relative to the end of the region. */
 		END
-	}
-
-	/** A description of a slice (range) of an IO object. */
-	final class Slice {
-		/** The index where the slice starts. */
-		public final Integer start;
-		/**
-		 * The index where the slice stops. (One after the last accessible
-		 * byte.)
-		 */
-		public final Integer stop;
-
-		private Slice(Integer start, Integer stop) {
-			this.start = start;
-			this.stop = stop;
-		}
-
-		/**
-		 * Create a new slice from the start position to the end of the IO
-		 * object.
-		 *
-		 * @param start
-		 *            Where to start.
-		 * @return The slice
-		 */
-		public static Slice from(int start) {
-			return new Slice(start, null);
-		}
-
-		/**
-		 * Create a new slice from the beginning to the stop position of the IO
-		 * object.
-		 *
-		 * @param end
-		 *            Where to finish.
-		 * @return The slice
-		 */
-		public static Slice to(int end) {
-			return new Slice(null, end);
-		}
-
-		/**
-		 * Create a new slice, from the the start position to the stop position,
-		 * of the IO object.
-		 *
-		 * @param start
-		 *            Where to start.
-		 * @param end
-		 *            Where to finish.
-		 * @return The slice
-		 */
-		public static Slice over(int start, int end) {
-			return new Slice(start, end);
-		}
 	}
 
 	/**
