+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/io/FileIO.java
@@ -10,6 +10,7 @@ import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
 
 import uk.ac.manchester.spinnaker.processes.FillProcess.DataType;
+import uk.ac.manchester.spinnaker.utils.Slice;
 
 /** A file input/output interface to match the MemoryIO interface. */
 public class FileIO implements AbstractIO {
