+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/io/MemoryIO.java
@@ -7,6 +7,7 @@ import uk.ac.manchester.spinnaker.machine.HasChipLocation;
 import uk.ac.manchester.spinnaker.processes.FillProcess.DataType;
 import uk.ac.manchester.spinnaker.processes.Process.Exception;
 import uk.ac.manchester.spinnaker.transceiver.Transceiver;
+import uk.ac.manchester.spinnaker.utils.Slice;
 
 /** A file-like object for reading and writing memory. */
 public class MemoryIO implements AbstractIO {
